Further organize file structure.
