+++ .eslintrc.js
@@ -10,7 +10,9 @@ module.exports = {
     ecmaVersion: 'latest',
   },
   rules: {
-    'no-unused-vars': ['error', { argsIgnorePattern: '(next)|(models)' }],
+    'no-unused-vars': [
+      'error',
+      { argsIgnorePattern: '(next)|(models)|(Sequelize)' },
+    ],
   },
-  ignorePatterns: ['/migrations/**/*', '/seeders/**/*'],
 };
