+++ .prettierignore
@@ -1,3 +1,3 @@
 .github/ISSUE_TEMPLATE
 docs/requirements.md
-views/layout.hbs
+server/views/layout.hbs
