+++ .sequelizerc
@@ -1,0 +1,8 @@
+const path = require('path');
+
+module.exports = {
+  config: path.resolve('server', 'config', 'database.json'),
+  'models-path': path.resolve('server', 'db', 'models'),
+  'seeders-path': path.resolve('server', 'db', 'seeders'),
+  'migrations-path': path.resolve('server', 'db', 'migrations'),
+};
