+++ server/routes/listing/view.js
