+++ server/views/index.hbs
